+++ grayc/src/stdlib/fmt.h
@@ -26,7 +26,7 @@
  *@module fmt
  *@group Output
  *@sig printf(format string, args [T])
- *@desc Prints a formatted string to stdout. Uses C-style format directives: %d (int), %f (float), %s (string), %b (bool), %c (char). Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Prints a formatted string to stdout. Uses C-style format directives: %d (i64), %f (f64), %s (string), %b (bool), %c (char). Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   fmt.printf("hello %s, you are %d years old\n", "alice", 30)
@@ -38,7 +38,7 @@
  *@module fmt
  *@group Output
  *@sig sprintf(format string, args [T]) -> string
- *@desc Returns a formatted string without printing it. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Returns a formatted string without printing it. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   mut s string = fmt.sprintf("x = %d", 42)
@@ -51,7 +51,7 @@
  *@module fmt
  *@group Output
  *@sig printfln(format string, args [T])
- *@desc Prints a formatted string to stdout with a trailing newline. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Prints a formatted string to stdout with a trailing newline. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   fmt.printfln("hello %s, you are %d years old", "alice", 30)
@@ -63,7 +63,7 @@
  *@module fmt
  *@group Output
  *@sig eprintf(format string, args [T])
- *@desc Prints a formatted string to stderr. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Prints a formatted string to stderr. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   fmt.eprintf("error: %s\n", "something went wrong")
@@ -75,7 +75,7 @@
  *@module fmt
  *@group Output
  *@sig eprintfln(format string, args [T])
- *@desc Prints a formatted string to stderr with a trailing newline. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Prints a formatted string to stderr with a trailing newline. Uses the same format directives as printf. Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   fmt.eprintfln("error: %s", "something went wrong")
@@ -87,7 +87,7 @@
  *@module fmt
  *@group Output
  *@sig sprintfln(format string, args [T]) -> string
- *@desc Returns a formatted string with a trailing newline. Uses the same format directives as sprintf. Pass one argument per format directive; each argument is independently int, uint, float, string, bool, or char. Composite types are rejected.
+ *@desc Returns a formatted string with a trailing newline. Uses the same format directives as sprintf. Pass one argument per format directive; each argument is independently i64, u64, f64, string, bool, or char. Composite types are rejected.
  *@example
  *   import @fmt
  *   mut s string = fmt.sprintfln("x = %d", 42)
@@ -99,7 +99,7 @@
 /*@man pad_left
  *@module fmt
  *@group Padding
- *@sig pad_left(s string, width int, ch char) -> string
+ *@sig pad_left(s string, width i64, ch char) -> string
  *@desc Returns s padded on the left with ch until the total length reaches width. Returns s unchanged if it is already at least width characters long.
  *@example
  *   import @fmt
@@ -111,7 +111,7 @@ GrayString gray_fmt_pad_left(GrayArena *arena, GrayString str, int64_t width, in
 /*@man pad_right
  *@module fmt
  *@group Padding
- *@sig pad_right(s string, width int, ch char) -> string
+ *@sig pad_right(s string, width i64, ch char) -> string
  *@desc Returns s padded on the right with ch until the total length reaches width. Returns s unchanged if it is already at least width characters long.
  *@example
  *   import @fmt
@@ -123,7 +123,7 @@ GrayString gray_fmt_pad_right(GrayArena *arena, GrayString str, int64_t width, i
 /*@man center
  *@module fmt
  *@group Padding
- *@sig center(s string, width int, ch char) -> string
+ *@sig center(s string, width i64, ch char) -> string
  *@desc Returns s centered within width, padded on both sides with ch. If padding is uneven, the extra character goes on the right.
  *@example
  *   import @fmt
@@ -135,7 +135,7 @@ GrayString gray_fmt_center(GrayArena *arena, GrayString str, int64_t width, int3
 /*@man int_to_hex
  *@module fmt
  *@group Number Formatting
- *@sig int_to_hex(n int) -> string
+ *@sig int_to_hex(n i64) -> string
  *@desc Returns the integer n formatted as a lowercase hexadecimal string with no "0x" prefix.
  *@example
  *   import @fmt
@@ -147,7 +147,7 @@ GrayString gray_fmt_int_to_hex(GrayArena *arena, int64_t value);
 /*@man int_to_binary
  *@module fmt
  *@group Number Formatting
- *@sig int_to_binary(n int) -> string
+ *@sig int_to_binary(n i64) -> string
  *@desc Returns the integer n formatted as a binary string with no "0b" prefix.
  *@example
  *   import @fmt
@@ -159,7 +159,7 @@ GrayString gray_fmt_int_to_binary(GrayArena *arena, int64_t value);
 /*@man int_to_octal
  *@module fmt
  *@group Number Formatting
- *@sig int_to_octal(n int) -> string
+ *@sig int_to_octal(n i64) -> string
  *@desc Returns the integer n formatted as an octal string with no "0o" prefix.
  *@example
  *   import @fmt
@@ -171,7 +171,7 @@ GrayString gray_fmt_int_to_octal(GrayArena *arena, int64_t value);
 /*@man float_fixed
  *@module fmt
  *@group Number Formatting
- *@sig float_fixed(f float, decimals int) -> string
+ *@sig float_fixed(f f64, decimals i64) -> string
  *@desc Returns f formatted with exactly decimals digits after the decimal point.
  *@example
  *   import @fmt
@@ -183,7 +183,7 @@ GrayString gray_fmt_float_fixed(GrayArena *arena, double value, int64_t decimals
 /*@man float_sci
  *@module fmt
  *@group Number Formatting
- *@sig float_sci(f float) -> string
+ *@sig float_sci(f f64) -> string
  *@desc Returns f formatted in scientific notation (e.g. "3.14e+00").
  *@example
  *   import @fmt
@@ -195,7 +195,7 @@ GrayString gray_fmt_float_sci(GrayArena *arena, double value);
 /*@man format_number
  *@module fmt
  *@group Number Formatting
- *@sig format_number(n int) -> string
+ *@sig format_number(n i64) -> string
  *@desc Returns n as a decimal string with ASCII comma thousands separators (no locale). 1234567 becomes "1,234,567"; -1000 becomes "-1,000".
  *@example
  *   import @fmt
@@ -207,7 +207,7 @@ GrayString gray_fmt_format_number(GrayArena *arena, int64_t value);
 /*@man format_bytes
  *@module fmt
  *@group Number Formatting
- *@sig format_bytes(n int) -> string
+ *@sig format_bytes(n i64) -> string
  *@desc Returns n formatted as a human-readable byte count using binary units (B, KiB, MiB, GiB, TiB, PiB). Values below 1024 are shown as whole bytes ("1023 B"); larger values use one decimal place ("1.5 KiB"). A negative n keeps a leading minus on the magnitude.
  *@example
  *   import @fmt
